fxies bug with editing scores

+++ FlipTrack/Views/GamesPlayedView.swift
@@ -1,56 +1,6 @@
 import SwiftUI
 
-struct ScoreEditView: View {
-    let game: Game
-    let scoreIndex: Int
-    let numberFormatter: NumberFormatter
-    let done: () -> Void
-    
-    @State private var editedScore = ""
-    @FocusState var focus: Bool
-    
-    var formatted: String {
-        let num = NSNumber(value: game.scores[scoreIndex])
-        return numberFormatter.string(from: num) ?? ""
-    }
-    
-    var body: some View {
-        VStack(alignment: .center) {
-            Text(formatted)
-                .bold()
-                .foregroundStyle(.red)
-                .padding(.top, 16)
-            TextField("Score", text: $editedScore)
-                .bold()
-                .focused($focus)
-                .keyboardType(.numberPad)
-                .multilineTextAlignment(.center)
-                .presentationDetents([.height(160)])
-                .presentationDragIndicator(.hidden)
-                .onAppear() {
-                    focus = true
-                }
-            Button(" Save ") {
-                if let newScore = Int(editedScore), newScore > 0, game.scores[scoreIndex] != newScore {
-                    game.scores[scoreIndex] = newScore
-                    try? game.modelContext?.save()
-                }
-                done()
-            }
-            .controlSize(.small)
-            .buttonStyle(.borderedProminent)
-            .padding(.top, 16)
-        }
-        .font(.title)
-        .onAppear {
-            editedScore = String(game.scores[scoreIndex])
-        }
-    }
-}
-
 struct GamesPlayedView: View {
-    @Environment(\.modelContext) private var context
-    
     let games: [Game]
     let formattedNumber: (Int) -> String
     let winningColor: (Game) -> Color
@@ -59,7 +9,6 @@ struct GamesPlayedView: View {
     @FocusState private var editFocus: Bool
     @State var editGame: Game? = nil
     @State var editScoreIndex = 0
-    @State var isEditing = false
     
     let numberFormatter = ({
         let formatter = NumberFormatter()
@@ -75,9 +24,8 @@ struct GamesPlayedView: View {
     ]}
     
     func startEdit(_ game: Game, _ scoreIndex: Int) {
-        editGame = game
         editScoreIndex = scoreIndex
-        isEditing = true
+        editGame = game
     }
     
     func textColor(_ game: Game, _ index: Int) -> Color {
@@ -92,7 +40,9 @@ struct GamesPlayedView: View {
         let hs = session.highScores
         return game.scores[index] == max(hs[0], hs[1])
     }
-
+    
+    var sortedGames: [Game] { games.sorted(using: SortDescriptor(\Game.nr)).reversed() }
+    
     var body: some View {
         VStack {
             Grid(alignment: .center) {
@@ -110,10 +60,7 @@ struct GamesPlayedView: View {
             ScrollView(.vertical) {
                 ScrollViewReader { scrollReader in
                     LazyVGrid(columns: columns) {
-                        //Text("#").padding(.trailing, 4)
-                        //Text("Andreas").foregroundColor(colorFor(0)).padding(.trailing, 4)
-                        //Text("Fredrik").foregroundColor(colorFor(1)).padding(.trailing, 10)
-                        ForEach(games.sorted(using: SortDescriptor(\Game.nr)).reversed()) { game in
+                        ForEach(sortedGames) { game in
                             Group {
                                 HStack {
                                     Spacer()
@@ -148,9 +95,11 @@ struct GamesPlayedView: View {
                             .gesture(
                                 LongPressGesture(minimumDuration: UIDevice.isSimulator ? 0.25 : 3)
                                     .onEnded { _ in
-                                        context.delete(game)
-                                        try! context.save()
-                                        scrollReader.scrollTo(0)
+                                        if let context = game.modelContext {
+                                            context.delete(game)
+                                            try! context.save()
+                                            scrollReader.scrollTo(0)
+                                        }
                                     }
                             )
                         }
@@ -162,12 +111,12 @@ struct GamesPlayedView: View {
             .scrollIndicators(.hidden)
             .padding(.bottom, 20)
         }
-        .sheet(isPresented: $isEditing) {
+        .sheet(item: $editGame) {
             ScoreEditView(
-                game: editGame ?? Game(),
+                game: $0,
                 scoreIndex: editScoreIndex,
                 numberFormatter: numberFormatter,
-                done: { isEditing = false }
+                done: { editGame = nil }
             )
         }
     }
@@ -178,7 +127,11 @@ struct GamesPlayedView: View {
         Game(nr: 1, scores: [12000, 32720], session: Session(date: Date.now )),
         Game(nr: 2, scores: [480230, 19260], session: Session(date: Date.now.addingTimeInterval(24 * 3600))),
     ]
-    GamesPlayedView(games: games, formattedNumber: { "\($0)" }, winningColor: { [Color.clear, Color.red, Color.green][$0.winningIndex + 1] }, colorFor: { [Color.red, Color.green][$0] })
+    GamesPlayedView(
+        games: games,
+        formattedNumber: { "\($0)" },
+        winningColor: { [Color.clear, Color.red, Color.green][$0.winningIndex + 1] },
+        colorFor: { [Color.red, Color.green][$0] }
+    )
     .preferredColorScheme(.dark)
-    
 }
